manually update the cache after resizing disk
